♻️ Move to components from page directory to view inside sources
