+++ pages/articles/index.ts
@@ -1,2 +1,2 @@
-import { Articles } from './articles'
+import { Articles } from '../../src/views/articles/articles'
 export default Articles
