+++ pages/home/index.ts
@@ -1,2 +1,2 @@
-import { Home } from './home'
+import { Home } from '../../src/views/home/home'
 export default Home
