+++ pages/resources/index.ts
@@ -1,2 +1,2 @@
-import { Resources } from './resources'
+import { Resources } from '../../src/views/resources/resources'
 export default Resources
