+++ src/views/post/post.module.scss
@@ -1,4 +1,5 @@
-@import '../variables';
+@import 'pages/variables';
+
 .article {
 
   display: grid;
