+++ src/views/resources/resources.module.scss
